requirements.txt
# RSV Timing / Beyfortus Analysis Pipeline Dependencies

# Data handling
pandas>=2.0
numpy>=1.24
pyarrow>=12.0

# API requests
requests>=2.28

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Geographic visualization
geopandas>=0.13

# Statistical functions
scipy>=1.10
